+++ Expanded Cheat Menu/Code/Menus/VehiclesMenu.lua
@@ -236,6 +236,20 @@ function OnMsg.ClassesGenerate()
 		OnAction = ChoGGi.MenuFuncs.SetDronesPerRCRover,
 	}
 
+	c = c + 1
+	Actions[c] = {ActionName = S[302535920001403--[[Drone Type--]]],
+		ActionMenubar = str_ExpandedCM_Drones,
+		ActionId = ".Drone Type",
+		ActionIcon = "CommonAssets/UI/Menu/UncollectObjects.tga",
+		RolloverText = function()
+			return ChoGGi.ComFuncs.SettingState(
+				GetMissionSponsor().drone_class or "Drone",
+				302535920001404--[[Change what type of drones will spawn (doesn't effect existing).--]]
+			)
+		end,
+		OnAction = ChoGGi.MenuFuncs.SetDroneType,
+	}
+
 	local str_ExpandedCM_Shuttles = "ECM.Expanded CM.Shuttles"
 	c = c + 1
 	Actions[c] = {ActionName = StringFormat("%s ..",S[745--[[Shuttles--]]]),