+++ Mods ChoGGi/Fix Rocket Stuck/metadata.lua
@@ -1,7 +1,7 @@
 return PlaceObj("ModDef", {
-	"title", "Fix: Rocket Stuck v0.3",
-	"version", 3,
-	"saved", 1542801600,
+	"title", "Fix: Rocket Stuck v0.4",
+	"version", 4,
+	"saved", 1542888000,
 	"image", "Preview.png",
 	"id", "ChoGGi_FixExpeditionRocketStuckUnloading",
 	"steam_id", "1567028510",
@@ -13,6 +13,9 @@ return PlaceObj("ModDef", {
 	"description", [[This will check on load game for certain rockets stuck on the ground unable to do jack.
 So far:
 Unloading colonists.
+Unloading colonist crew from expedition.
 Maintenance */5.
-Drones stuck inside.]],
+Drones stuck inside.
+
+If this doesn't fix it for you, then I'll need a copy of your saved game.]],
 })