+++ Mods ChoGGi/Map Images Pack/metadata.lua
@@ -1,0 +1,14 @@
+return PlaceObj("ModDef", {
+	"title", "Map Images Pack v0.1",
+	"version", 1,
+	"saved", 1542888000,
+	"image", "Preview.png",
+	"id", "ChoGGi_MapImagesPack",
+	"steam_id", "1571465108",
+	"author", "ChoGGi",
+--~ 	"code", {},
+	"lua_revision", LuaRevision,
+	"description", [[These are used in more than one of my mods, so no sense in having dupes.
+This mod doesn't need to be enabled, just installed.
+]],
+})